backend/app/requirements.txt
fastapi[standard]
pyjwt
passlib[bcrypt]
python-multipart
alembic
pydantic
pydantic_settings
pymysql
